requirements.txt
#django
django~=3.0.0
django-appdata>=0.3.0
django-filer>=1.4.2
django-parler
django-sortedm2m
django-taggit>=0.23.0
django-treebeard>=4.0,<5.0

#django debug
django-debug-toolbar>=1.11

#django cms
django-cms~=3.7.3
djangocms-admin-style>=1.5,<1.6
djangocms-text-ckeditor>=3.7,<4.0
djangocms-file>=2.3,<2.5
djangocms-link>=2.5,<2.7
djangocms-icon>=1.4,<1.6
djangocms-picture>=2.3,<2.5
djangocms-bootstrap4>=1.5,<1.7
djangocms-style>=2.2,<2.4
djangocms-snippet>=2.2,<2.4
djangocms-googlemap>=1.3,<1.5
djangocms-video>=2.1,<2.4
djangocms-audio>=1.1.0
djangocms_history>=1.0.0
aldryn-newsblog>=2.2.1
easy_thumbnails
#djangocms_column>=1.9

#gsoc irc requirements
fredirc==0.3.0

#gsoc pdf requirements
pdfminer.six==20181108
chardet==3.0.4
phonenumbers==8.10.6

#gsoc comments requirements
profanityfilter>=2.0.6

#git pre-commit hook
pre-commit>=1.14.4

#google api for calander
google-api-python-client>=1.7.9
google-auth-httplib2>=0.0.3
google-auth-oauthlib>=0.3.0

#github api for pushing html pages
PyGithub>=1.43.7

bleach>=3.1.0
beautifulsoup4>=4.8.0

django-simple-cookie-consent>=0.1.1

mysqlclient>=1.4.4
importlib-resources>=1.0.2

<<<<<<< HEAD
html5lib>=1.0.1
Pillow>=3.0
six
pytz
=======
backport_collections==0.1
html5lib>=1.0.1
Pillow>=3.0
lxml
six
pytz
babel>=2.8.0
>>>>>>> 86b094f... Further work on Django 3 migration
